✨ RENDERER: Refactor Media Discovery
💡 What: Refactored `findAllMedia` logic into a shared `dom-scripts.ts` utility.
🎯 Why: To eliminate code duplication across `CdpTimeDriver`, `SeekTimeDriver`, and `dom-scanner`.
📊 Impact: Ensures consistent media discovery behavior across all strategies and improves maintainability.
🔬 Verification: Verified logic equivalence; ran test suite.
