+++ packages/renderer/src/drivers/SeekTimeDriver.ts
@@ -1,6 +1,7 @@
 import { Page } from 'playwright';
 import { TimeDriver } from './TimeDriver.js';
 import { RANDOM_SEED_SCRIPT } from '../utils/random-seed.js';
+import { FIND_ALL_MEDIA_FUNCTION } from '../utils/dom-scripts.js';
 
 export class SeekTimeDriver implements TimeDriver {
   constructor(private timeout: number = 30000) {}
@@ -132,28 +133,7 @@ export class SeekTimeDriver implements TimeDriver {
         }
 
         // Helper to find all media elements, including in Shadow DOM
-        function findAllMedia(rootNode) {
-          const media = [];
-          // Check rootNode itself (if it is an Element)
-          if (rootNode.nodeType === Node.ELEMENT_NODE) {
-            const tagName = rootNode.tagName;
-            if (tagName === 'AUDIO' || tagName === 'VIDEO') {
-              media.push(rootNode);
-            }
-          }
-
-          const walker = document.createTreeWalker(rootNode, NodeFilter.SHOW_ELEMENT);
-          while (walker.nextNode()) {
-            const node = walker.currentNode;
-            if (node.tagName === 'AUDIO' || node.tagName === 'VIDEO') {
-              media.push(node);
-            }
-            if (node.shadowRoot) {
-              media.push.apply(media, findAllMedia(node.shadowRoot));
-            }
-          }
-          return media;
-        }
+        ${FIND_ALL_MEDIA_FUNCTION}
 
         // 2. Synchronize media elements (video, audio)
         const mediaElements = findAllMedia(document);
